refactor: consolidate api services and stores
