+++ client/src/api/agent-cycle.service.ts
