+++ client/src/api/index.ts
@@ -1,7 +1,5 @@
-
-
-
-
-
-
-export { default as apiClient } from '@/api/apiClient'
+export { default as apiClient } from '@/api/apiClient';
+export * as dashboardService from '@/api/dashboardService';
+export * as toolService from '@/api/toolService';
+export * as workflowService from '@/api/workflowService';
+export { AgentCycleService } from '@/api/agent-cycle.service';
