+++ client/src/features/agents/hooks/useAgentSelectors.ts
@@ -1,5 +1,5 @@
 import { AgentDTO } from '@/features/agents/services/agentService'
-import { useAgentStore } from '@/store/agentStore'
+import { useAgentStore } from '@/features/agents/store/agentStore'
 
 export const useAgentsList = (): AgentDTO[] =>
   useAgentStore((state) => state.agents as unknown as AgentDTO[])
