+++ client/src/features/chat/components/ConversationList.tsx
@@ -11,7 +11,7 @@ import {
 } from '@/components/ui/dropdown-menu'
 import { ScrollArea } from '@/components/ui/scroll-area'
 import { cn } from '@/lib/utils'
-import { useChatStore } from '@/store/chatStore'
+import { useChatStore } from '@/features/chat/store/chatStore'
 
 const ConversationList: React.FC = () => {
   const {
