+++ client/src/features/chat/store/sessionStore.ts
