+++ client/src/features/dashboard/components/MeusAgentesCard.tsx
@@ -6,7 +6,7 @@ import { Badge } from '@/components/ui/badge'
 import { Button } from '@/components/ui/button'
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import { ScrollArea } from '@/components/ui/scroll-area'
-import type { Agent } from '@/store/dashboardStore'
+import type { Agent } from '@/features/dashboard/store/dashboardStore'
 
 interface MeusAgentesCardProps {
   agents: Agent[]
