+++ client/src/features/roi-dashboard/routes/RoiDashboardPage.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useState } from 'react';
 import KpiCard from '@/features/dashboard/components/KpiCard';
 import RoiChart from '@/features/dashboard/components/RoiChart';
 import { ComponentSkeleton } from '@/components/ui/ComponentSkeleton';
-import { dashboardService } from '@/services/dashboardService';
+import { dashboardService } from '@/api/dashboardService';
 import { RoiMetricsResponse, RoiTimeSeriesResponse } from '@/types/common';
 
 const RoiDashboardPage: React.FC = () => {
